+++ setup.cfg
@@ -75,13 +75,13 @@ console_scripts =
 diracx =
 	properties_module = diracx.core:properties
 diracx.db.sql =
-	AuthDB = diracx.db:AuthDB
-	JobDB = diracx.db:JobDB
-	JobLoggingDB = diracx.db:JobLoggingDB
-	SandboxMetadataDB = diracx.db:SandboxMetadataDB
+	AuthDB = diracx.db.sql:AuthDB
+	JobDB = diracx.db.sql:JobDB
+	JobLoggingDB = diracx.db.sql:JobLoggingDB
+	SandboxMetadataDB = diracx.db.sql:SandboxMetadataDB
 	#DummyDB = diracx.db:DummyDB
 diracx.db.os =
-	JobParametersDB = diracx.db:JobParametersDB
+	JobParametersDB = diracx.db.os:JobParametersDB
 diracx.services =
 	jobs = diracx.routers.job_manager:router
 	config = diracx.routers.configuration:router
