+++ src/diracx/db/__init__.py
@@ -1,7 +1,5 @@
-__all__ = ("AuthDB", "JobDB", "JobLoggingDB", "SandboxMetadataDB")
+from __future__ import annotations
 
-from .sql.auth.db import AuthDB
-from .sql.jobs.db import JobDB, JobLoggingDB
-from .sql.sandbox_metadata.db import SandboxMetadataDB
+__all__ = ("sql", "os")
 
-# from .dummy.db import DummyDB
+from . import os, sql
