+++ src/diracx/db/sql/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+__all__ = ("AuthDB", "JobDB", "JobLoggingDB", "SandboxMetadataDB")
+
+from .auth.db import AuthDB
+from .jobs.db import JobDB, JobLoggingDB
+from .sandbox_metadata.db import SandboxMetadataDB
